Commented a bit of code, preparing to PID tune arm
some rambling space for my idea for the arm:
since the pid loop will make the arm go quickly towards the desired position, if it goes up, the arm will make the robot throw itself over with a  quick stop of the arm, a part of the PID loop. what i'm thinking about doing is using the little slider on the joystick to calculate the required motor output at a lot of different angles and creating a sine function to calculate different holding outputs for every single angle.  With this, I can do a bang-bang closed loop by repeated calculating the holding output and adding a constant to that output. It will be slow, but that makes sure it's safe and doesn't cause the robot to tip over. It's also much easier to tune and will help me keep my sanity. It's also probably gonna be accurate as long as the constant added to the holding output is relatively low. Another random thing is that I might want to add a safeguard so that if the NavX detects the robot tilting too much (the absolute value of the pitch crosses a threshold value), the robot will bring the arm down in an attempt to counterbalance the tilt. I saw in one of the regional livestreams an elevator bot tipped over when the elevator was extended out. The bot probably wouldn't have tipped over if the driver reacted quickly and brought the arm down in order to balance the robot.